pyproject.toml
[project]
name = "math-agent-vikhr"
version = "0.1.0"
description = "Multi-agent system for automated theorem proving using Lean 4"
requires-python = ">=3.10"
dependencies = [
    "smolagents[toolkit]>=1.21.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "requests>=2.28.0",
    "tenacity>=8.0.0",
    "brotli>=1.1.0",
    "PyYAML>=6.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "arize-phoenix>=11.32.1",
    "openinference-instrumentation-smolagents>=0.1.16",
    "openinference-instrumentation-openai>=0.1.31",
]
